backend/scraper/requirements.txt
youtube-transcript-api==0.6.3
Flask==2.3.2
Flask_Cors==4.0.0
pytesseract==0.3.13
# pillow==11.1.0
beautifulsoup4==4.12.3
lxml-html-clean==0.4.1
newspaper3k==0.2.8
nltk==3.9.1
flask_swagger_ui==4.11.1
flask-restx==1.3.0
Werkzeug==3.0.1
selenium
gunicorn
requests